requirements.txt
# Core
python-dotenv==1.0.0

# Google API
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0

# Requests
requests==2.31.0

# Logging
python-dateutil==2.8.2
